Struct expander::Expander

pub struct Expander { /* private fields */ }

Expander to replace a tokenstream by a include to a file

Implementations

impl Expander

pub fn new(filename: impl AsRef<str>) -> Self

Create a new expander.

pub fn add_comment(self, comment: impl Into<Option<String>>) -> Self

Add a header comment.

pub fn fmt(self, edition: impl Into<Edition>) -> Self

Format the resulting file, for readability.

pub fn dry(self, dry: bool) -> Self

Do not modify the provided tokenstream.

pub fn verbose(self, verbose: bool) -> Self

Print the path of the generated file to stderr during the proc-macro invocation.

pub fn maybe_write_to_out_dir( self, tokens: impl Into<Result<TokenStream, Error>> ) -> Result<Result<TokenStream>, Error>

Create a file with filename under env!("OUT_DIR") if it’s not an Err(_).

pub fn write_to_out_dir(self, tokens: TokenStream) -> Result<TokenStream, Error>

Create a file with filename under env!("OUT_DIR").

pub fn maybe_write_to( self, maybe_tokens: impl Into<Result<TokenStream, Error>>, dest_dir: &Path ) -> Result<Result<TokenStream>, Error>

Create a file with filename at dest if it’s not an Err(_).

pub fn write_to( self, tokens: TokenStream, dest_dir: &Path ) -> Result<TokenStream, Error>

Create a file with self.filename in dest_dir.